Moved the Analytics page
